requirements.txt
python-dotenv~=1.0.1
starlette~=0.38.2
Flask~=3.0.3
fastapi~=0.112.2
injector~=0.22.0
h11~=0.14.0
pip~=24.2
Jinja2~=3.1.4
typing_extensions~=4.12.2
sniffio~=1.3.1
MarkupSafe~=2.1.5
Werkzeug~=3.0.4
click~=8.1.7
blinker~=1.8.2
itsdangerous~=2.2.0
pydantic~=2.8.2
pydantic_core~=2.20.1
anyio~=4.4.0
uvicorn~=0.30.6
annotated-types~=0.7.0
setuptools~=65.5.0